agent_builder/requirements.txt
fastapi
uvicorn
ibm-watsonx-ai
python-dotenv